Add cop for BlockLength
This cop discriminates against blocks that are too long (by default 25
lines). It also does not like blocks inside specs, which leads to lots
of complaints in RSpec tests.
The cop takes an argument to ignore certain methods, so I provided an
array with the common RSpec methods `["describe", "context", "it"]`.